@@ -81,7 +81,7 @@
     "chai-as-promised": "^7.1.1",
     "chromedriver": "^86.0.0",
     "concurrently": "^5.3.0",
-    "eslint": "^7.11.0",
+    "eslint": "^7.12.1",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-plugin-import": "^2.22.1",
