@@ -1,5 +1,3 @@
-'use strict';
-
 module.exports = function (grunt) {
   // To-Do check what we need and add/remove as needed...
   [
@@ -13,13 +11,10 @@ module.exports = function (grunt) {
     'grunt-mocha-test', // Server side test runner
     'grunt-bower-task',
     'grunt-gitinfo',
-    'grunt-karma' // Client side test runner.
+    'grunt-karma', // Client side test runner.
   ].forEach(grunt.loadNpmTasks);
 
   grunt.loadTasks('tasks');
-
-  // To create HTML test files from a template. To-Do: Dunno if this is even needed or not
-  var TEST_BASE_DIR = 'test/';
 
   grunt.initConfig({
     terser: {
@@ -29,8 +24,8 @@ module.exports = function (grunt) {
           safari10: true,
           ecma: 2016,
           sourceMap: {
-            includeSources: true
-          }
+            includeSources: true,
+          },
         },
         files: [
           {
@@ -38,15 +33,15 @@ module.exports = function (grunt) {
             src: '*.js',
             dest: './web/js/min',
             cwd: './web/js',
-            ext: '.min.js'
-          }
-        ]
+            ext: '.min.js',
+          },
+        ],
       },
       prod_build: {
         options: {
           compress: true,
           safari10: true,
-          ecma: 2016
+          ecma: 2016,
         },
         files: [
           {
@@ -54,18 +49,18 @@ module.exports = function (grunt) {
             src: '*.js',
             dest: './web/js/min',
             cwd: './web/js',
-            ext: '.min.js'
-          }
-        ]
-      }
+            ext: '.min.js',
+          },
+        ],
+      },
 
     },
     concat: {
       dist: {
         options: {
-          process: function (src, filepath) {
-            return src + '\n //# sourceMappingURL= ' + filepath.substring(filepath.lastIndexOf('/') + 1) + '.map';
-          }
+          process(src, filepath) {
+            return `${src}\n //# sourceMappingURL= ${filepath.substring(filepath.lastIndexOf('/') + 1)}.map`;
+          },
         },
         files: {
           './web/js/min/chatView.min.js': ['./web/js/min/chatView.min.js'],
@@ -80,9 +75,9 @@ module.exports = function (grunt) {
           './web/js/min/recordingsController.min.js': ['./web/js/min/recordingsController.min.js'],
           './web/js/min/screenShareView.min.js': ['./web/js/min/screenShareView.min.js'],
           './web/js/min/screenShareController.min.js': ['./web/js/min/screenShareController.min.js'],
-          './web/js/min/roomController.min.js': ['./web/js/min/roomController.min.js']
-        }
-      }
+          './web/js/min/roomController.min.js': ['./web/js/min/roomController.min.js'],
+        },
+      },
     },
     mochaTest: {
       unit: {
@@ -90,21 +85,21 @@ module.exports = function (grunt) {
           reporter: 'spec',
           captureFile: 'resultsUnit.txt',
           quiet: false,
-          clearRequireCache: true
+          clearRequireCache: true,
         },
         src: [
-          'test/server/**/*_spec.js'
-        ]
+          'test/server/**/*_spec.js',
+        ],
       },
       rest: {
         options: {
           reporter: 'spec',
           captureFile: 'resultsRest.txt',
           quiet: false,
-          clearRequireCache: true
+          clearRequireCache: true,
         },
-        src: ['test/api/**/*_spec.js']
-      }
+        src: ['test/api/**/*_spec.js'],
+      },
     },
 
     bower: {
@@ -117,19 +112,19 @@ module.exports = function (grunt) {
           verbose: false,
           cleanTargetDir: false,
           cleanBowerDir: true,
-          bowerOptions: {}
-        }
-      }
+          bowerOptions: {},
+        },
+      },
     },
 
     karma: {
       options: {
-        configFile: 'karma.conf.js'
+        configFile: 'karma.conf.js',
       },
       integration: {
-        singleRun: true
+        singleRun: true,
       },
-      dev: {}
+      dev: {},
     },
 
     clean: {
@@ -138,8 +133,8 @@ module.exports = function (grunt) {
 
     gitinfo: {
       options: {
-        cwd: '.'
-      }
+        cwd: '.',
+      },
     },
 
     less: {
@@ -147,16 +142,16 @@ module.exports = function (grunt) {
         options: {
           compress: true,
           yuicompress: true,
-          optimization: 2
+          optimization: 2,
         },
         files: {
           'web/css/landing.opentok.css': 'web/less/landing.less',
           'web/css/room.opentok.css': 'web/less/room.less',
           'web/css/endMeeting.opentok.css': 'web/less/endMeeting.less',
           'web/css/annotation.opentok.css': 'web/less/annotation.less',
-          'web/css/hangoutScroll.css': 'web/less/hangoutScroll.less'
-        }
-      }
+          'web/css/hangoutScroll.css': 'web/less/hangoutScroll.less',
+        },
+      },
     },
     critical: {
       test: {
@@ -164,21 +159,21 @@ module.exports = function (grunt) {
           base: './',
           css: ['web/css/landing.opentok.css'],
           target: {
-            uncritical: 'web/css/landing-uncritical.opentok.css'
+            uncritical: 'web/css/landing-uncritical.opentok.css',
           },
-          minify: true
+          minify: true,
         },
         src: 'views/index.ejs',
-        dest: 'web/css/landing-critical.opentok.css'
-      }
+        dest: 'web/css/landing-critical.opentok.css',
+      },
     },
     autoprefixer: {
       options: {
-        browsers: ['last 5 versions']
+        browsers: ['last 5 versions'],
       },
       dist: {
-        src: 'web/css/*.css'
-      }
+        src: 'web/css/*.css',
+      },
     },
 
     watch: {
@@ -186,23 +181,23 @@ module.exports = function (grunt) {
         files: ['./web/**/*.less'],
         tasks: ['less', 'autoprefixer'],
         options: {
-          nospawn: false
-        }
+          nospawn: false,
+        },
       },
       server: {
         options: {
-          nospawn: false
+          nospawn: false,
         },
         files: ['./server.js', 'server/**/*.js', 'test/server/**/*.js'],
-        tasks: ['serverTest']
-      }
-    }
+        tasks: ['serverTest'],
+      },
+    },
   });
 
   // On watch events, if the changed file is a test file then configure mochaTest to only
   // run the tests from that file. Otherwise run all the tests
-  var defaultTestSrc = grunt.config('mochaTest.unit.src');
-  grunt.event.on('watch', function (action, filepath) {
+  const defaultTestSrc = grunt.config('mochaTest.unit.src');
+  grunt.event.on('watch', (action, filepath) => {
     grunt.config('mochaTest.unit.src', defaultTestSrc);
     if (filepath.match('test/')) {
       grunt.config('mochaTest.unit.src', filepath);
@@ -214,44 +209,44 @@ module.exports = function (grunt) {
     'autoprefixer',
     'terser:pages',
     'concat',
-    'critical'
+    'critical',
   ]);
 
   grunt.registerTask('clientBuild-Prod', 'Build css files', [
     'less',
     'autoprefixer',
     'terser:prod_build',
-    'critical'
+    'critical',
   ]);
 
   grunt.registerTask('clientDev', 'Watch for changes on less files', [
     'clientBuild',
-    'watch'
+    'watch',
   ]);
 
   grunt.registerTask(
     'clientTest',
     'Launch client unit tests in shell with Karma + PhantomJS',
-    ['karma:dev']
+    ['karma:dev'],
   );
 
   grunt.registerTask('precommit', 'Run precommit tests', [
     'karma:integration',
     'mochaTest:unit',
-    'apiTest'
+    'apiTest',
   ]);
 
   grunt.registerTask('serverTest', 'Launch server unit tests', [
-    'mochaTest:unit'
+    'mochaTest:unit',
   ]);
 
   grunt.registerTask('apiTest', 'Launch server unit tests', ['mochaTest:rest']);
 
   grunt.registerTask('archivesTest', 'Launch server unit tests', [
-    'mochaTest:archives'
+    'mochaTest:archives',
   ]);
 
-  grunt.registerTask('test', 'Launch server unit tests', function () {
+  grunt.registerTask('test', 'Launch server unit tests', () => {
     grunt.task.run(['configTests', 'serverTest']);
     if (grunt.option('enable-archives-test')) {
       grunt.task.run(['archivesTest']);
@@ -262,7 +257,7 @@ module.exports = function (grunt) {
   grunt.registerTask('configTests', [
     'preBowerInstall',
     'bower:install',
-    'postBowerInstall'
+    'postBowerInstall',
   ]);
 
   grunt.registerTask('initialConfig', ['clientBuild', 'configTests']);
