@@ -9,11 +9,9 @@ module.exports = function (config) {
     // base path that will be used to resolve all patterns (eg. files, exclude)
     basePath: '',
 
-
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
     frameworks: ['mocha', 'sinon', 'chai', 'browserify'],
-
 
     // list of files / patterns to load in the browser
     files: [
@@ -34,9 +32,8 @@ module.exports = function (config) {
       'web/js/layouts.js',
       'web/**/*.js',
       'test/unit/**/*.html',
-      'test/unit/**/*.js'
+      'test/unit/**/*.js',
     ],
-
 
     // list of files to exclude
     exclude: [
@@ -46,9 +43,8 @@ module.exports = function (config) {
       'web/js/vendor/lazy_loader.js',
       'web/js/helpers/OTHelper.js',
       'web/js/rtcApp.js',
-      'web/js/vendor/opentok-annotation.js'
+      'web/js/vendor/opentok-annotation.js',
     ],
-
 
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
@@ -65,29 +61,25 @@ module.exports = function (config) {
       'test/unit/itemsHandler_spec.js': ['browserify'],
       'test/unit/layoutMenuView_spec.js': ['browserify'],
       'test/unit/layouts_spec.js': ['browserify'],
-      'test/unit/precallController_spec.js': ['browserify']
+      'test/unit/precallController_spec.js': ['browserify'],
 
     },
-
 
     // test results reporter to use
     // possible values: 'dots', 'progress'
     // available reporters: https://npmjs.org/browse/keyword/karma-reporter
     reporters: ['progress', 'coverage'],
 
-
     // web server port
     port: 9876,
-
 
     // enable / disable colors in the output (reporters and logs)
     colors: true,
 
-
     // level of logging
-    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
+    // possible values: config.LOG_DISABLE || config.LOG_ERROR ||
+    // config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
     logLevel: config.LOG_INFO,
-
 
     // enable / disable watching file and executing tests whenever any file changes
     autoWatch: false,
@@ -95,13 +87,13 @@ module.exports = function (config) {
     coverageReporter: {
       dir: 'coverage',
       instrumenter: {
-        'web/**/*.js': ['istanbul']
+        'web/**/*.js': ['istanbul'],
       },
       reporters: [
         { type: 'html', subdir: 'report-html' },
         { type: 'lcov', subdir: 'report-lcov' },
-        { type: 'lcovonly', subdir: '.', file: 'report-lcovonly.txt' }
-      ]
+        { type: 'lcovonly', subdir: '.', file: 'report-lcovonly.txt' },
+      ],
     },
 
     // start these browsers
@@ -111,13 +103,13 @@ module.exports = function (config) {
     customLaunchers: {
       Chrome_travis_ci: {
         base: 'Chrome',
-        flags: ['--no-sandbox']
-      }
+        flags: ['--no-sandbox'],
+      },
     },
 
     // Continuous Integration mode
     // if true, Karma captures browsers, runs the tests and exits
-    singleRun: true
+    singleRun: true,
   };
 
   if (process.env.TRAVIS) {
@@ -125,5 +117,4 @@ module.exports = function (config) {
   }
 
   config.set(configuration);
-
 };
