@@ -2,25 +2,18 @@
 // defined on the api.yaml swagger 2.0 file
 // Usage:
 // node server -h
+const { Server } = require('swagger-boilerplate');
 
-const SwaggerBP = require('swagger-boilerplate');
-const Utils = SwaggerBP.Utils;
-const Logger = Utils.MultiLevelLogger;
-const logger = new Logger('HTTP Server App');
-
-const Server = require('swagger-boilerplate').Server;
-
-const server =
- new Server({
-    apiFile: './api.yml',
-    modulePath: `${__dirname}/server/`,
-    appName: 'OpenTokRTC Main',
-    staticOptions: {
-      dotfiles: 'ignore',
-      extensions: ['jpg'],
-      index: false,
-      redirect: false
-    }
-  });
+const server = new Server({
+  apiFile: './api.yml',
+  modulePath: `${__dirname}/server/`,
+  appName: 'OpenTokRTC Main',
+  staticOptions: {
+    dotfiles: 'ignore',
+    extensions: ['jpg'],
+    index: false,
+    redirect: false,
+  },
+});
 
 server.start();
