@@ -1,10 +1,10 @@
-'use strict';
+const { env } = process;
+const SwaggerBP = require('swagger-boilerplate');
 
-var env = process.env;
-var SwaggerBP = require('swagger-boilerplate');
-var Utils = SwaggerBP.Utils;
-var Logger = Utils.MultiLevelLogger;
-const Redis = require("ioredis");
+const { Utils } = SwaggerBP;
+const Logger = Utils.MultiLevelLogger;
+const Redis = require('ioredis');
+
 const redis = new Redis(env.REDIS_URL || env.REDISTOGO_URL || ''); // uses defaults unless given configuration object
 
 class ArchiveLocalStorage {
@@ -35,8 +35,9 @@ class ArchiveLocalStorage {
           return this.logger.log('Get archives error:', error);
         }
         return false;
-      });
-  };
+      },
+    );
+  }
 
   async updateArchive(aArchive) {
     const stringSessionInfo = await redis.get(this.roomNameKey);
@@ -44,7 +45,7 @@ class ArchiveLocalStorage {
     if (!sessionInfo.archives) sessionInfo.archives = {};
     sessionInfo.archives[aArchive.id] = aArchive;
     await redis.set(this.roomNameKey, JSON.stringify(sessionInfo));
-    this.sendBroadcastSignal(sessionInfo.archives);  
+    this.sendBroadcastSignal(sessionInfo.archives);
   }
 
   async removeArchive(aArchiveId) {
@@ -52,9 +53,8 @@ class ArchiveLocalStorage {
     const sessionInfo = JSON.parse(stringSessionInfo);
     delete sessionInfo.archives[aArchiveId];
     await redis.set(this.roomNameKey, JSON.stringify(sessionInfo));
-    this.sendBroadcastSignal(sessionInfo.archives);  
+    this.sendBroadcastSignal(sessionInfo.archives);
   }
-
 }
-  
+
 module.exports = ArchiveLocalStorage;
