@@ -1,13 +1,10 @@
-'use strict';
+const _ = require('lodash');
+const { Utils } = require('swagger-boilerplate');
 
-var _ = require('lodash');
-var Utils = require('swagger-boilerplate').Utils;
-
-var readFile = Utils.promisify(require('fs').readFile);
+const readFile = Utils.promisify(require('fs').readFile);
 const defaultJsonConfigPath = require('./serverConstants').DEFAULT_JSON_CONFIG_PATH;
 
-var env = process.env;
-var exports = module.exports = {};
+const { env } = process;
 
 class Config {
   constructor(aConfigJson) {
@@ -21,7 +18,7 @@ class Config {
   // parser: Function to return value as correct type i.e. parseInt.
   // required: boolean, whether value can be undefined, or throw error.
   get(aParam) {
-    var value = env[aParam.envVar] || _.get(this.configJson, aParam.jsonPath, aParam.defaultValue);
+    let value = env[aParam.envVar] || _.get(this.configJson, aParam.jsonPath, aParam.defaultValue);
     if (aParam.required && value === undefined) {
       throw new Error(`${aParam.envVar} must be supplied.`);
     }
@@ -35,7 +32,7 @@ class Config {
 // Promise wrapper around reading config file.
 // Returns a Config instance
 exports.readConfigJson = () => {
-  var configPath = env.OTDEMO_CONFIG_FILE_PATH || defaultJsonConfigPath;
+  let configPath = env.OTDEMO_CONFIG_FILE_PATH || defaultJsonConfigPath;
   return new Promise((resolve, reject) => {
     try {
       configPath = require.resolve(configPath);
@@ -43,5 +40,5 @@ exports.readConfigJson = () => {
     } catch (e) {
       (e.code === 'MODULE_NOT_FOUND' && (resolve('{}') || true)) || reject(e);
     }
-  }).then(data => new Config(JSON.parse(data)));
+  }).then((data) => new Config(JSON.parse(data)));
 };
