@@ -1,7 +1,5 @@
-'use strict';
-
 const { OAuth2Client } = require('google-auth-library');
-const Utils = require('swagger-boilerplate').Utils;
+const { Utils } = require('swagger-boilerplate');
 
 /* eslint-disable class-methods-use-this */
 class DisabledGoogleAuthStategy {
@@ -33,12 +31,12 @@ class EnabledGoogleAuthStrategy {
           }
           resolve();
         })
-        .catch(err => reject(err));
+        .catch((err) => reject(err));
     });
   }
 }
 
 module.exports = {
   DisabledGoogleAuthStategy,
-  EnabledGoogleAuthStrategy
+  EnabledGoogleAuthStrategy,
 };
