@@ -1,13 +1,11 @@
 // Just to have all the useful constants on a single place.
 
-'use strict';
-
-var E = module.exports;
+const E = module.exports;
 
 // json config will be boolean but if environment override will be string
 function parseBool(input) {
-  return input === true ||
-  (typeof input === 'string' && input.toLowerCase() === 'true');
+  return input === true
+  || (typeof input === 'string' && input.toLowerCase() === 'true');
 }
 
 E.DEFAULT_JSON_CONFIG_PATH = '../config/config.json';
@@ -23,7 +21,7 @@ E.DEFAULT_USER_NAME = 'Anonymous User';
 E.REDIS_ROOM_PREFIX = 'otrtc_room__';
 E.REDIS_PHONE_PREFIX = 'otrtc_phone_';
 
-E.REDIS_ROOM_MATCHES = E.REDIS_ROOM_PREFIX + '*';
+E.REDIS_ROOM_MATCHES = `${E.REDIS_ROOM_PREFIX}*`;
 
 E.OPENTOK_API_KEY = { envVar: 'TB_API_KEY', jsonPath: 'OpenTok.apiKey', required: true };
 
@@ -33,34 +31,56 @@ E.OPENTOK_PRECALL_API_KEY = { envVar: 'TB_PRECALL_API_KEY', jsonPath: 'precallTe
 
 E.OPENTOK_PRECALL_API_SECRET = { envVar: 'TB_PRECALL_API_SECRET', jsonPath: 'precallTest.apiSecret' };
 
-E.OPENTOK_JS_URL = { envVar: 'TB_JS_URL', jsonPath: 'OpenTok.jsUrl', required: false, defaultValue: 'https://static.opentok.com/v2/js/opentok.min.js' };
+E.OPENTOK_JS_URL = {
+  envVar: 'TB_JS_URL', jsonPath: 'OpenTok.jsUrl', required: false, defaultValue: 'https://static.opentok.com/v2/js/opentok.min.js',
+};
 
 // Sessions should not live forever. So we'll store the last time a session was used and if when
 // we fetch it from Redis we determine it's older than this max age (in days). This is the key
 // where that value (in days) should be stored. By default, sessions live two days.
-E.OPENTOK_MAX_SESSION_AGE = { envVar: 'TB_MAX_SESSION_AGE', jsonPath: 'OpenTok.maxSessionAge', defaultValue: 2, parser: parseFloat };
+E.OPENTOK_MAX_SESSION_AGE = {
+  envVar: 'TB_MAX_SESSION_AGE', jsonPath: 'OpenTok.maxSessionAge', defaultValue: 2, parser: parseFloat,
+};
 
-E.ENABLE_ARCHIVING = { envVar: 'ENABLE_ARCHIVING', jsonPath: 'Archiving.enabled', defaultValue: true, parser: parseBool };
+E.ENABLE_ARCHIVING = {
+  envVar: 'ENABLE_ARCHIVING', jsonPath: 'Archiving.enabled', defaultValue: true, parser: parseBool,
+};
 
-E.ARCHIVE_ALWAYS = { envVar: 'ARCHIVE_ALWAYS', jsonPath: 'Archiving.archiveAlways', defaultValue: false, parser: parseBool };
+E.ARCHIVE_ALWAYS = {
+  envVar: 'ARCHIVE_ALWAYS', jsonPath: 'Archiving.archiveAlways', defaultValue: false, parser: parseBool,
+};
 
 // Timeout (in milliseconds) for polling for archive status change updates. Set this to zero
 // to disable polling. This is the initial timeout (timeout for the first poll).
-E.ARCHIVE_POLLING_INITIAL_TIMEOUT = { envVar: 'ARCHIVE_TIMEOUT', jsonPath: 'Archiving.pollingInitialTimeout', defaultValue: 5000, parser: parseInt };
+E.ARCHIVE_POLLING_INITIAL_TIMEOUT = {
+  envVar: 'ARCHIVE_TIMEOUT', jsonPath: 'Archiving.pollingInitialTimeout', defaultValue: 5000, parser: parseInt,
+};
 
 // Timeout multiplier. After the first poll (if it fails) the next one will apply this multiplier
 // successively. Set to a lower number to poll often.
-E.ARCHIVE_POLLING_TIMEOUT_MULTIPLIER = { envVar: 'TIMEOUT_MULTIPLIER', jsonPath: 'Archiving.pollingTimeoutMultiplier', defaultValue: 1.5, parser: parseFloat };
+E.ARCHIVE_POLLING_TIMEOUT_MULTIPLIER = {
+  envVar: 'TIMEOUT_MULTIPLIER', jsonPath: 'Archiving.pollingTimeoutMultiplier', defaultValue: 1.5, parser: parseFloat,
+};
 
-E.ENABLE_ARCHIVE_MANAGER = { envVar: 'ENABLE_ARCHIVE_MANAGER', jsonPath: 'Archiving.archiveManager.enabled', defaultValue: false, parser: parseBool };
+E.ENABLE_ARCHIVE_MANAGER = {
+  envVar: 'ENABLE_ARCHIVE_MANAGER', jsonPath: 'Archiving.archiveManager.enabled', defaultValue: false, parser: parseBool,
+};
 
-E.ENABLE_MUTE_ALL = { envVar: 'ENABLE_MUTE_ALL', jsonPath: 'enableMuteAll', defaultValue: true, parser: parseBool };
+E.ENABLE_MUTE_ALL = {
+  envVar: 'ENABLE_MUTE_ALL', jsonPath: 'enableMuteAll', defaultValue: true, parser: parseBool,
+};
 
-E.ENABLE_EMOJI = { envVar: 'ENABLE_EMOJI', jsonPath: 'enableEmoji', defaultValue: false, parser: parseBool  };
+E.ENABLE_EMOJI = {
+  envVar: 'ENABLE_EMOJI', jsonPath: 'enableEmoji', defaultValue: false, parser: parseBool,
+};
 
-E.ENABLE_STOP_RECEIVING_VIDEO = { envVar: 'ENABLE_STOP_RECEIVING_VIDEO', jsonPath: 'enableStopReceivingVideo', defaultValue: true, parser: parseBool };
+E.ENABLE_STOP_RECEIVING_VIDEO = {
+  envVar: 'ENABLE_STOP_RECEIVING_VIDEO', jsonPath: 'enableStopReceivingVideo', defaultValue: true, parser: parseBool,
+};
 
-E.MAX_USERS_PER_ROOM = { envVar: 'MAX_USERS_PER_ROOM', jsonPath: 'maxUsersPerRoom', defaultValue: 0, parser: parseInt };
+E.MAX_USERS_PER_ROOM = {
+  envVar: 'MAX_USERS_PER_ROOM', jsonPath: 'maxUsersPerRoom', defaultValue: 0, parser: parseInt,
+};
 
 E.FEEDBACK_URL = { envVar: 'FEEDBACK_URL', jsonPath: 'Feedback.url', defaultValue: '' };
 
@@ -70,24 +90,40 @@ E.HOTJAR_ID = { envVar: 'HOTJAR_ID', jsonPath: 'Feedback.hotjarId', defaultValue
 
 E.HOTJAR_VERSION = { envVar: 'HOTJAR_VERSION', jsonPath: 'Feedback.hotjarVersion', defaultValue: '' };
 
-E.ENABLE_FEEDBACK = { envVar: 'ENABLE_FEEDBACK', jsonPath: 'Feedback.enabled', defaultValue: false, parser: parseBool };
+E.ENABLE_FEEDBACK = {
+  envVar: 'ENABLE_FEEDBACK', jsonPath: 'Feedback.enabled', defaultValue: false, parser: parseBool,
+};
 
-E.ENABLE_SCREENSHARING = { envVar: 'ENABLE_SCREENSHARING', jsonPath: 'Screensharing.enabled', defaultValue: false, parser: parseBool };
+E.ENABLE_SCREENSHARING = {
+  envVar: 'ENABLE_SCREENSHARING', jsonPath: 'Screensharing.enabled', defaultValue: false, parser: parseBool,
+};
 
-E.ENABLE_PRECALL_TEST = { envVar: 'ENABLE_PRECALL_TEST', jsonPath: 'precallTest.enabled', defaultValue: true, parser: parseBool };
+E.ENABLE_PRECALL_TEST = {
+  envVar: 'ENABLE_PRECALL_TEST', jsonPath: 'precallTest.enabled', defaultValue: true, parser: parseBool,
+};
 
-E.USE_GOOGLE_FONTS = { envVar: 'USE_GOOGLE_FONTS', jsonPath: 'useGoogleFonts', required: false, defaultValue: true, parser: parseBool };
+E.USE_GOOGLE_FONTS = {
+  envVar: 'USE_GOOGLE_FONTS', jsonPath: 'useGoogleFonts', required: false, defaultValue: true, parser: parseBool,
+};
 
-E.JQUERY_URL = { envVar: 'JQUERY_URL', jsonPath: 'jqueryUrl', required: false, defaultValue: 'https://ajax.googleapis.com/ajax/libs/jquery' };
+E.JQUERY_URL = {
+  envVar: 'JQUERY_URL', jsonPath: 'jqueryUrl', required: false, defaultValue: 'https://ajax.googleapis.com/ajax/libs/jquery',
+};
 
 // Chrome AddOn extension Id for sharing screen
 E.CHROME_EXTENSION_ID = { envVar: 'CHROME_EXTENSION_ID', jsonPath: 'Screensharing.chromeExtensionId' };
 
-E.ENABLE_ANNOTATIONS = { envVar: 'ENABLE_ANNOTATIONS', jsonPath: 'Screensharing.annotations.enabled', defaultValue: true, parser: parseBool };
+E.ENABLE_ANNOTATIONS = {
+  envVar: 'ENABLE_ANNOTATIONS', jsonPath: 'Screensharing.annotations.enabled', defaultValue: true, parser: parseBool,
+};
 
-E.ENABLE_ROOM_LOCKING = { envVar: 'ENABLE_ROOM_LOCKING', jsonPath: 'enableRoomLocking', defaultValue: true, parser: parseBool };
+E.ENABLE_ROOM_LOCKING = {
+  envVar: 'ENABLE_ROOM_LOCKING', jsonPath: 'enableRoomLocking', defaultValue: true, parser: parseBool,
+};
 
-E.AUTO_GENERATE_ROOM_NAME = { envVar: 'AUTO_GENERATE_ROOM_NAME', jsonPath: 'autoGenerateRoomName', defaultValue: true, parser: parseBool };
+E.AUTO_GENERATE_ROOM_NAME = {
+  envVar: 'AUTO_GENERATE_ROOM_NAME', jsonPath: 'autoGenerateRoomName', defaultValue: true, parser: parseBool,
+};
 
 // Do we want to allow being used inside an iframe?
 // This can be:
@@ -97,7 +133,6 @@ E.AUTO_GENERATE_ROOM_NAME = { envVar: 'AUTO_GENERATE_ROOM_NAME', jsonPath: 'auto
 //  'sameorigin': Set X-Frame-Options to 'SAMEORIGIN'
 // We don't allow restricting it to some URIs because it doesn't work on Chrome
 E.ALLOW_IFRAMING = { envVar: 'ALLOW_IFRAMING', jsonPath: 'allowIframing', defaultValue: 'never' };
-
 
 E.DEFAULT_TEMPLATE = { envVar: 'DEFAULT_TEMPLATE', jsonPath: 'defaultTemplate', defaultValue: 'room' };
 
@@ -109,7 +144,9 @@ E.IOS_URL_PREFIX = { envVar: 'IOS_URL_PREFIX', jsonPath: 'IOSUrlPrefix', default
 
 E.DEFAULT_INDEX_PAGE = { envVar: 'DEFAULT_INDEX_PAGE', jsonPath: 'defaultPageIndex', defaultValue: undefined };
 
-E.SHOW_TOS = { envVar: 'SHOW_TOS', jsonPath: 'showTos', defaultValue: false, parser: parseBool };
+E.SHOW_TOS = {
+  envVar: 'SHOW_TOS', jsonPath: 'showTos', defaultValue: false, parser: parseBool,
+};
 
 E.MEETINGS_RATE_PER_MINUTE = { envVar: 'MEETINGS_RATE_PER_MINUTE', jsonPath: 'meetingsRatePerMinute', defaultValue: -1 };
 
@@ -117,7 +154,9 @@ E.MIN_MEETING_NAME_LENGTH = { envVar: 'MIN_MEETING_NAME_LENGTH', jsonPath: 'minM
 
 E.PUBLISHER_RESOLUTION = { envVar: 'PUBLISHER_RESOLUTION', jsonPath: 'OpenTok.publisherResolution', defaultValue: '1280x720' };
 
-E.SIP_ENABLED = { envVar: 'SIP_ENABLED', jsonPath: 'SIP.enabled', defaultValue: false, parser: parseBool };
+E.SIP_ENABLED = {
+  envVar: 'SIP_ENABLED', jsonPath: 'SIP.enabled', defaultValue: false, parser: parseBool,
+};
 
 E.SIP_URI = { envVar: 'SIP_URI', jsonPath: 'SIP.uri', defaultValue: '' };
 
@@ -125,7 +164,9 @@ E.SIP_USERNAME = { envVar: 'SIP_USERNAME', jsonPath: 'SIP.username', defaultValu
 
 E.SIP_PASSWORD = { envVar: 'SIP_PASSWORD', jsonPath: 'SIP.password', defaultValue: '' };
 
-E.SIP_REQUIRE_GOOGLE_AUTH = { envVar: 'SIP_REQUIRE_GOOGLE_AUTH', jsonPath: 'SIP.requireGoogleAuth', defaultValue: false, parser: parseBool };
+E.SIP_REQUIRE_GOOGLE_AUTH = {
+  envVar: 'SIP_REQUIRE_GOOGLE_AUTH', jsonPath: 'SIP.requireGoogleAuth', defaultValue: false, parser: parseBool,
+};
 
 E.GOOGLE_CLIENT_ID = { envVar: 'GOOGLE_CLIENT_ID', jsonPath: 'Google.clientId', defaultValue: '' };
 
