@@ -1,7 +1,7 @@
 var chai = require('chai');
 var request = require('supertest');
 
-var expect = chai.expect;
+var { expect } = chai;
 
 const TEST_LOG_LEVEL = 0;
 
@@ -26,16 +26,14 @@ describe('OpenTokRTC server', () => {
     // So that's what '.' is. OTOH, the requires are relative to *this* file.
     // Yep, I don't like that either. Nope, I can't do anything about that.
     var YAML = require('yamljs');
-    var loadYAML =
-         apiFile => new Promise((resolve, reject) => {
-           try {
-             YAML.load(apiFile, resolve);
-           } catch (e) {
-             reject(e);
-           }
-         });
+    var loadYAML = (apiFile) => new Promise((resolve, reject) => {
+      try {
+        YAML.load(apiFile, resolve);
+      } catch (e) {
+        reject(e);
+      }
+    });
     loadYAML('./api.yml').then((apiSpec) => {
-
       app = (require('swagger-boilerplate').App)({
         modulePath: __dirname + '/../../server/', // eslint-disable-line no-path-concat
         staticPath: '../../web',
@@ -57,8 +55,8 @@ describe('OpenTokRTC server', () => {
     var aObject = aRes.body;
     for (var i = 0, l = aAttributes.length; i < l; i++) {
       if (!aObject[aAttributes[i]]) {
-        throw new Error('Missing required attribute: ' + aAttributes[i] +
-                        ' in ' + JSON.stringify(aObject));
+        throw new Error('Missing required attribute: ' + aAttributes[i]
+                        + ' in ' + JSON.stringify(aObject));
       }
     }
     return undefined;
@@ -126,7 +124,6 @@ describe('OpenTokRTC server', () => {
       .expect('Content-Type', new RegExp('text/html'))
       .expect(200, done);
   });
-
 
   it('GET /room/:roomName?template=unknownTemplate without auth, return default', (done) => {
     request(app)
