@@ -1,8 +1,9 @@
 var sinonTest = require('sinon-test');
+
 var test = sinonTest(sinon);
 sinon.test = test;
-var assert = chai.assert;
-var expect = chai.expect;
+var { assert } = chai;
+var { expect } = chai;
 var should = chai.should();
 
 describe('FeedbackController', () => {
@@ -29,7 +30,7 @@ describe('FeedbackController', () => {
         logEvent() {},
       },
     };
-    sinon.stub(LazyLoader, 'load').callsFake(  resources => Promise.resolve());
+    sinon.stub(LazyLoader, 'load').callsFake((resources) => Promise.resolve());
   });
 
   after(() => {
