@@ -1,4 +1,4 @@
-var expect = chai.expect;
+var { expect } = chai;
 
 describe('roomStatus', () => {
   var expectedHandlers = ['signal:status', 'signal:statusACK', 'connectionCreated',
@@ -33,65 +33,65 @@ describe('roomStatus', () => {
     });
 
     it('should initialize properly the object and return the handlers set',
-       sinon.test(() => {
-         var handlers = [];
+      sinon.test(() => {
+        var handlers = [];
 
-         RoomStatus.init(handlers);
+        RoomStatus.init(handlers);
 
-         expect(handlers.length).to.be.equals(1);
-         var statusHandlers = handlers[0];
-         expect(expectedHandlers.every(elem => statusHandlers[elem] !== undefined)).to.be.true;
-       }));
+        expect(handlers.length).to.be.equals(1);
+        var statusHandlers = handlers[0];
+        expect(expectedHandlers.every((elem) => statusHandlers[elem] !== undefined)).to.be.true;
+      }));
   });
 
   describe('#handlers', () => {
     describe('#signal:status', () => {
-      it('should populate room\'s status, send ack, remove the listener and ' +
-         'send updatedRemotely event', sinon.test(function (done) {
-           var status = {
-             chat: [{
-               sender: 'aSender1',
-               time: Utils.getCurrentTime(),
-               text: 'a text 1',
-             }, {
-               sender: 'aSender2',
-               time: Utils.getCurrentTime(),
-               text: 'a text 2',
-             }],
-             key2: {},
-           };
+      it('should populate room\'s status, send ack, remove the listener and '
+         + 'send updatedRemotely event', sinon.test(function (done) {
+        var status = {
+          chat: [{
+            sender: 'aSender1',
+            time: Utils.getCurrentTime(),
+            text: 'a text 1',
+          }, {
+            sender: 'aSender2',
+            time: Utils.getCurrentTime(),
+            text: 'a text 2',
+          }],
+          key2: {},
+        };
 
-           var signalEvt = {
-             data: JSON.stringify(status),
-             from: {
-               connectionId: 'otherUser',
-             },
-             type: 'status',
-           };
+        var signalEvt = {
+          data: JSON.stringify(status),
+          from: {
+            connectionId: 'otherUser',
+          },
+          type: 'status',
+        };
 
-           this.spy(otHelper, 'sendSignal');
-           this.spy(otHelper, 'removeListener');
+        this.spy(otHelper, 'sendSignal');
+        this.spy(otHelper, 'removeListener');
 
         // Testing that roomStatus:updatedRemotely was called we just put
         // a listener over it, if it wasn't executed correctly we'll have a
         // time out.
-           window.addEventListener('roomStatus:updatedRemotely', (evt) => {
-             done();
-           });
+        window.addEventListener('roomStatus:updatedRemotely', (evt) => {
+          done();
+        });
 
-           var handlers = [];
-           RoomStatus.init(handlers);
+        var handlers = [];
+        RoomStatus.init(handlers);
 
-           var hndls = OTHelper.bindHandlers(handlers[0]);
-           hndls['signal:status'](signalEvt);
+        var hndls = OTHelper.bindHandlers(handlers[0]);
+        hndls['signal:status'](signalEvt);
 
-           expect(OTHelper.sendSignal.calledWith('statusACK')).to.be.true;
+        expect(OTHelper.sendSignal.calledWith('statusACK')).to.be.true;
 
-           expect(OTHelper.removeListener.calledWith('signal:status')).to.be.true;
-           Object.keys(status).forEach((key) => {
-             expect(RoomStatus.get(key)).to.be.deep.equal(status[key]);
-           });
-         }));
+        expect(OTHelper.removeListener.calledWith('signal:status')).to.be.true;
+        Object.keys(status).forEach((key) => {
+          expect(RoomStatus.get(key)).to.be.deep.equal(status[key]);
+        });
+      }));
     });
 
     describe('#connectionCreated', () => {
@@ -121,46 +121,46 @@ describe('roomStatus', () => {
       }
 
       it('should send status when a different user connects and I was the oldest connected one',
-         sinon.test(function () {
-           this.spy(OTHelper, 'sendSignal');
-           var otherCreationTime = myCreationTime + 1;
-           verifySend('otherConnId', 'otherUser', otherCreationTime);
-         }));
+        sinon.test(function () {
+          this.spy(OTHelper, 'sendSignal');
+          var otherCreationTime = myCreationTime + 1;
+          verifySend('otherConnId', 'otherUser', otherCreationTime);
+        }));
 
-      it('should send status when a different user with my same identifier connects and ' +
-         ' I was the oldest connected', sinon.test(function () {
-           this.spy(OTHelper, 'sendSignal');
-           var otherCreationTime = myCreationTime + 1;
-           verifySend('otherConnId', 'mySelf', otherCreationTime);
-         }));
+      it('should send status when a different user with my same identifier connects and '
+         + ' I was the oldest connected', sinon.test(function () {
+        this.spy(OTHelper, 'sendSignal');
+        var otherCreationTime = myCreationTime + 1;
+        verifySend('otherConnId', 'mySelf', otherCreationTime);
+      }));
 
-      it('should not send the status when a different user connects and I was not the oldest ' +
-         'connected one', sinon.test(function () {
-           this.spy(OTHelper, 'sendSignal');
+      it('should not send the status when a different user connects and I was not the oldest '
+         + 'connected one', sinon.test(function () {
+        this.spy(OTHelper, 'sendSignal');
 
-           var otherConnectedCreationTime = myCreationTime - 1;
-           var newConnectedCreationTime = myCreationTime + 1;
+        var otherConnectedCreationTime = myCreationTime - 1;
+        var newConnectedCreationTime = myCreationTime + 1;
 
-           statusHndls.connectionCreated(getSignalEvent('connIdfirstUsr',
-                                                        otherConnectedCreationTime,
-                                                        'firstUser'));
-           statusHndls.connectionCreated(getSignalEvent('connNewUsr',
-                                                        newConnectedCreationTime),
-                                                        'newUser');
+        statusHndls.connectionCreated(getSignalEvent('connIdfirstUsr',
+          otherConnectedCreationTime,
+          'firstUser'));
+        statusHndls.connectionCreated(getSignalEvent('connNewUsr',
+          newConnectedCreationTime),
+        'newUser');
 
-           expect(OTHelper.sendSignal.callCount).to.be.equal(0);
-         }));
+        expect(OTHelper.sendSignal.callCount).to.be.equal(0);
+      }));
 
       it('should not send the history when I receive a connect event for myself',
-         sinon.test(function () {
-           this.spy(OTHelper, 'sendSignal');
+        sinon.test(function () {
+          this.spy(OTHelper, 'sendSignal');
 
-           statusHndls.connectionCreated(getSignalEvent('myConnectionId',
-                                                        myCreationTime,
-                                                        'mySelf'));
+          statusHndls.connectionCreated(getSignalEvent('myConnectionId',
+            myCreationTime,
+            'mySelf'));
 
-           expect(OTHelper.sendSignal.callCount).to.be.equal(0);
-         }));
+          expect(OTHelper.sendSignal.callCount).to.be.equal(0);
+        }));
     });
   });
 });
