@@ -32,10 +32,9 @@
  *
  */
 
-!(global => {
-  const transEndEventName =
-    ('WebkitTransition' in document.documentElement.style) ?
-      'webkitTransitionEnd' : 'transitionend';
+!((global) => {
+  const transEndEventName = ('WebkitTransition' in document.documentElement.style)
+    ? 'webkitTransitionEnd' : 'transitionend';
 
   const HORIZONTAL_OFFSET = 10;
   const VERTICAL_OFFSET = 4;
@@ -45,11 +44,11 @@
   /*
    * Closes all bubbles clicking outside them
    */
-  const onBodyClicked = evt => {
+  const onBodyClicked = (evt) => {
     document.body.removeEventListener('click', onBodyClicked);
-    Object.keys(bubbles).forEach(id => {
+    Object.keys(bubbles).forEach((id) => {
       const bubble = bubbles[id];
-      let target = evt.target;
+      let { target } = evt;
       if (bubble.associatedWith !== target) {
         // pointer-events is not working on IE so we can receive as target a child of
         // "change layout" item in main menu
@@ -77,7 +76,7 @@
     this._onHidden = this._onHidden.bind(this);
 
     // Bubbles consumes 'click' events in order not to be closed automatically
-    this.container.addEventListener('click', e => {
+    this.container.addEventListener('click', (e) => {
       e.stopImmediatePropagation();
     });
   };
@@ -86,23 +85,22 @@
     show() {
       const bubble = this;
 
-      this.bubbleShown =
-        this.bubbleShown || new Promise(resolve => {
-          const container = bubble.container;
+      this.bubbleShown = this.bubbleShown || new Promise((resolve) => {
+        const { container } = bubble;
 
-          container.removeEventListener(transEndEventName, bubble._onHidden);
-          container.addEventListener(transEndEventName, bubble._onShown);
-          container.addEventListener(transEndEventName, function onEnd() {
-            container.removeEventListener(transEndEventName, onEnd);
-            resolve();
-          });
-
-          bubble._takePlace();
-          bubble._visible = true;
-          setTimeout(() => {
-            container.classList.add('show');
-          }, 50); // Give the chance to paint the UI element before fading in
+        container.removeEventListener(transEndEventName, bubble._onHidden);
+        container.addEventListener(transEndEventName, bubble._onShown);
+        container.addEventListener(transEndEventName, function onEnd() {
+          container.removeEventListener(transEndEventName, onEnd);
+          resolve();
         });
+
+        bubble._takePlace();
+        bubble._visible = true;
+        setTimeout(() => {
+          container.classList.add('show');
+        }, 50); // Give the chance to paint the UI element before fading in
+      });
 
       return this.bubbleShown;
     },
@@ -110,22 +108,22 @@
     hide() {
       const bubble = this;
 
-      this.bubbleHidden =
-        this.bubbleHidden || new Promise(resolve => {
-          const container = bubble.container;
+      this.bubbleHidden = this.bubbleHidden || new Promise((resolve) => {
+        const { container } = bubble;
 
-          container.removeEventListener(transEndEventName, bubble._onShown);
-          container.addEventListener(transEndEventName, bubble._onHidden);
-          container.addEventListener(transEndEventName, function onEnd() {
-            container.removeEventListener(transEndEventName, onEnd);
-            bubble.bubbleShown = bubble.bubbleHidden = null;
-            resolve();
-          });
-
-          setTimeout(() => {
-            container.classList.remove('show');
-          }, 50); // Give the chance to paint the UI element before fading out
+        container.removeEventListener(transEndEventName, bubble._onShown);
+        container.addEventListener(transEndEventName, bubble._onHidden);
+        container.addEventListener(transEndEventName, function onEnd() {
+          container.removeEventListener(transEndEventName, onEnd);
+          bubble.bubbleShown = null;
+          bubble.bubbleHidden = null;
+          resolve();
         });
+
+        setTimeout(() => {
+          container.classList.remove('show');
+        }, 50); // Give the chance to paint the UI element before fading out
+      });
 
       return this.bubbleHidden;
     },
@@ -145,13 +143,13 @@
     },
 
     set _visible(value) {
-      const classList = this.container.classList;
+      const { classList } = this.container;
       value ? classList.add('visible') : classList.remove('visible');
     },
 
     _takePlace() {
       const rectObject = this.associatedWith.getBoundingClientRect();
-      const container = this.container;
+      const { container } = this;
       if (this.topArrow) {
         container.style.right = `${window.innerWidth - rectObject.right + 25}px`;
         container.style.top = `${rectObject.bottom + VERTICAL_OFFSET}px`;
@@ -159,7 +157,7 @@
         container.style.left = `${rectObject.right + HORIZONTAL_OFFSET}px`;
         container.style.top = `${rectObject.top - rectObject.height}px`;
       }
-    }
+    },
   };
 
   global.BubbleFactory = {
@@ -172,10 +170,11 @@
       let instance = bubbles[id];
 
       if (!instance) {
-        instance = bubbles[id] = new Bubble(id);
+        instance = new Bubble(id);
+        bubbles[id] = new Bubble(id);
       }
 
       return instance;
-    }
+    },
   };
 })(this);
