@@ -1,6 +1,6 @@
 /* global OT, FeedbackView */
 
-!(global => {
+!((global) => {
   let otHelper;
 
   const eventHandlers = {
@@ -16,7 +16,7 @@
         videoScore: report.videoScore,
         description: report.description,
         clientSystemTime: new Date().getTime(),
-        source: document.location.href
+        source: document.location.href,
       };
       const xhr = new XMLHttpRequest();
       const url = window.feedbackUrl;
@@ -29,27 +29,25 @@
         partnerId: otHelper.session.apiKey,
         sessionId: otHelper.session.id,
         connectionId: otHelper.session.connection.id,
-        publisherId: otHelper.publisherId
+        publisherId: otHelper.publisherId,
       };
       OT.reportIssue((error, reportId) => {
         if (!error) {
           loggedEvent.reportIssueId = reportId;
         }
       });
-    }
+    },
   };
 
-  const init = (aOTHelper, aReportIssueLevel) => {
-    return LazyLoader.load([
-      '/js/min/feedbackView.min.js'
-    ]).then(() => {
-      otHelper = aOTHelper;
-      Utils.addEventsHandlers('feedbackView:', eventHandlers, global);
-      FeedbackView.init(aReportIssueLevel);
-    });
-  };
+  const init = (aOTHelper, aReportIssueLevel) => LazyLoader.load([
+    '/js/min/feedbackView.min.js',
+  ]).then(() => {
+    otHelper = aOTHelper;
+    Utils.addEventsHandlers('feedbackView:', eventHandlers, global);
+    FeedbackView.init(aReportIssueLevel);
+  });
 
   global.FeedbackController = {
-    init
+    init,
   };
 })(this);
