@@ -1,11 +1,10 @@
-/* global Utils, Request, RoomStatus, RoomView, LayoutManager, LazyLoader, Modal,
+/* global Utils, RoomStatus, RoomView, LazyLoader, Modal,
 ChatController, GoogleAuth, LayoutMenuController, OTHelper, PrecallController,
 RecordingsController, ScreenShareController, FeedbackController,
 PhoneNumberController, ResizeSensor, maxUsersPerRoom */
 
-!(exports => {
-  const debug =
-    new Utils.MultiLevelLogger('roomController.js', Utils.MultiLevelLogger.DEFAULT_LEVELS.all);
+!((exports) => {
+  const debug = new Utils.MultiLevelLogger('roomController.js', Utils.MultiLevelLogger.DEFAULT_LEVELS.all);
 
   let otHelper;
   let numUsrsInRoom = 0;
@@ -18,19 +17,17 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
   let googleAuth = null;
 
   let setPublisherReady;
-  const publisherReady = new Promise(resolve => {
+  const publisherReady = new Promise((resolve) => {
     setPublisherReady = resolve;
   });
 
   const STATUS_KEY = 'room';
   let _sharedStatus = {
-    roomMuted: false
+    roomMuted: false,
   };
 
-  let userName = window.userName;
+  let { userName } = window;
   let roomURI = null;
-  let resolutionAlgorithm = null;
-  let debugPreferredResolution = null;
   let token = null;
 
   const publisherOptions = {
@@ -44,8 +41,8 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
       buttonDisplayMode: 'off',
       nameDisplayMode: 'off',
       videoDisabledDisplayMode: 'on',
-      showArchiveStatus: false
-    }
+      showArchiveStatus: false,
+    },
   };
 
   const subscriberOptions = {
@@ -59,8 +56,8 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         buttonDisplayMode: 'off',
         nameDisplayMode: 'off',
         videoDisabledDisplayMode: 'auto',
-        showArchiveStatus: false
-      }
+        showArchiveStatus: false,
+      },
     },
     screen: {
       height: '100%',
@@ -71,8 +68,8 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         audioLevelDisplayMode: 'off',
         buttonDisplayMode: 'off',
         nameDisplayMode: 'on',
-        videoDisabledDisplayMode: 'off'
-      }
+        videoDisabledDisplayMode: 'off',
+      },
     },
     noVideo: {
       height: '100%',
@@ -83,12 +80,12 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         audioLevelDisplayMode: 'auto',
         buttonDisplayMode: 'off',
         nameDisplayMode: 'on',
-        videoDisabledDisplayMode: 'off'
-      }
-    }
+        videoDisabledDisplayMode: 'off',
+      },
+    },
   };
 
-  const isMobile = () => { return typeof window.orientation !== 'undefined'; };
+  const isMobile = () => typeof window.orientation !== 'undefined';
 
   const SubscriberButtons = (streamVideoType, phoneNumber) => {
     const isScreenSharing = streamVideoType === 'screen';
@@ -102,7 +99,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         eventName: 'click',
         context: 'otHelper',
         action: 'toggleSubscribersVideo',
-        enabled: true
+        enabled: true,
       };
     }
 
@@ -113,7 +110,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         eventName: 'click',
         context: 'otHelper',
         action: 'toggleSubscribersAudio',
-        enabled: true
+        enabled: true,
       };
     }
     if (phoneNumber && (phoneNumber in dialedNumberTokens)) {
@@ -123,7 +120,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         eventName: 'click',
         context: 'otHelper',
         action: 'hangup',
-        enabled: true
+        enabled: true,
       };
     }
 
@@ -137,7 +134,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
       eventName: 'click',
       context: 'otHelper',
       action: 'togglePublisherVideo',
-      enabled: true
+      enabled: true,
     },
     audio: {
       eventFiredName: 'roomView:buttonClick',
@@ -145,97 +142,38 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
       eventName: 'click',
       context: 'otHelper',
       action: 'togglePublisherAudio',
-      enabled: true
-    }
+      enabled: true,
+    },
   };
 
   let subscriberStreams = { };
-  var dialedNumberTokens = {};
+  const dialedNumberTokens = { };
 
-  // We want to use media priorization on the subscriber streams. We're going to restrict the
-  // maximum width and height to the one that's actually displayed. To do that, we're going to
-  // observe changes on the elements that hold the subscribers.
-  // Note that mutationObserver only works on IE11+, but that the previous alternative doesn't
-  // work all that well either.
-  const processMutation = aMutation => {
-    const elem = aMutation.target;
-    if ((aMutation.attributeName !== 'style' && aMutation.attributeName !== 'class') ||
-        elem.data('streamType') !== 'camera') {
-      return;
-    }
-    const streamId = elem.data('id');
-    const subscriberPromise =
-      subscriberStreams[streamId] && subscriberStreams[streamId].subscriberPromise;
-
-    subscriberPromise.then(subscriber => {
-      if (debugPreferredResolution) {
-        // If the user requested debugging this, we're going to export all the information through
-        // window so he can examine the values.
-        window.subscriberElem = window.subscriberElem || {};
-        window.subscriberElem[streamId] = elem;
-        window.subscriber = window.subscriber || {};
-        window.subscriber[streamId] = subscriber;
-        window.dumpResolutionInfo = window.dumpResolutionInfo || (() => {
-          Object.keys(window.subscriber)
-            .forEach(aSub => {
-              const sub = window.subscriber[aSub];
-              const stream = sub && sub.stream;
-              const vd = stream && stream.videoDimensions;
-              const streamPref = (stream && stream.getPreferredResolution()) ||
-                                 { width: 'NA', height: 'NA' };
-              stream && console.log( // eslint-disable-line no-console
-                'StreamId:', aSub, 'Real:', sub.videoWidth(), 'x', sub.videoHeight(),
-                'Stream.getPreferredResolution:', streamPref.width, 'x', streamPref.height,
-                'Stream.VDimension:', vd.width, 'x', vd.height
-              );
-            });
-        });
-      }
-
-      const parent = elem.parentNode;
-
-      const parentDimension = {
-        width: parent.clientWidth,
-        height: parent.clientHeight
-      };
-      const subsDimension = {
-        width: elem.clientWidth,
-        height: elem.clientHeight
-      };
-      otHelper.setPreferredResolution(subscriber, parentDimension, subsDimension, numUsrsInRoom - 1,
-        resolutionAlgorithm);
-    });
-  };
-  const _mutationObserver = exports.MutationObserver &&
-    new exports.MutationObserver(aMutations => {
-      aMutations.forEach(processMutation);
-    });
-
-  const sendVideoEvent = stream => {
+  const sendVideoEvent = (stream) => {
     if (!stream) {
       return;
     }
 
     Utils.sendEvent(`roomController:${stream.hasVideo ? 'videoEnabled' : 'videoDisabled'}`, {
-      id: stream.streamId
+      id: stream.streamId,
     });
   };
 
-  const sendArchivingOperation = operation => {
+  const sendArchivingOperation = (operation) => {
     const data = {
       userName,
       roomName: roomURI,
-      operation
+      operation,
     };
 
     Request.sendArchivingOperation(data);
   };
 
-  const dialOut = phoneNumber => {
+  const dialOut = (phoneNumber) => {
     const alreadyInCall = Object.keys(subscriberStreams)
-      .some(streamId => {
+      .some((streamId) => {
         if (subscriberStreams[streamId]) {
-          const stream = subscriberStreams[streamId].stream;
+          const { stream } = subscriberStreams[streamId];
           return (stream.isSip && stream.name === phoneNumber);
         }
         return false;
@@ -253,22 +191,22 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
       }
       const data = {
         phoneNumber,
-        googleIdToken
+        googleIdToken,
       };
       Request.dialOut(roomURI, data);
       dialedNumberTokens[phoneNumber] = googleIdToken;
     }
   };
 
-  const hangup = streamId => {
+  const hangup = (streamId) => {
     if (!subscriberStreams[streamId]) {
       return;
     }
-    const stream = subscriberStreams[streamId].stream;
+    const { stream } = subscriberStreams[streamId];
     if (!stream.isSip) {
       return;
     }
-    const phoneNumber = stream.phoneNumber;
+    const { phoneNumber } = stream;
     if (!(phoneNumber in dialedNumberTokens)) {
       return;
     }
@@ -281,32 +219,32 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
     updatedRemotely() {
       publisherReady.then(() => {
         _sharedStatus = RoomStatus.get(STATUS_KEY);
-        const roomMuted = _sharedStatus.roomMuted;
+        const { roomMuted } = _sharedStatus;
         setAudioStatus(roomMuted);
         roomMuted && Utils.sendEvent('roomController:roomMuted', { isJoining: true });
       });
-    }
+    },
   };
 
   const changeSubscriberStatus = (name, status) => {
     _disabledAllVideos = status;
 
-    Object.keys(subscriberStreams).forEach(aStreamId => {
-      if (subscriberStreams[aStreamId] &&
-          subscriberStreams[aStreamId].stream.videoType === 'camera') {
+    Object.keys(subscriberStreams).forEach((aStreamId) => {
+      if (subscriberStreams[aStreamId]
+          && subscriberStreams[aStreamId].stream.videoType === 'camera') {
         pushSubscriberButton(aStreamId, name, status);
       }
     });
   };
 
-  var pushSubscriberButton = (streamId, name, status) => {
+  const pushSubscriberButton = (streamId, name, status) => {
     viewEventHandlers.buttonClick({
       detail: {
         streamId,
         name,
         disableAll: true,
-        status
-      }
+        status,
+      },
     });
   };
 
@@ -318,12 +256,11 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
     otHelper.sendSignal('roomLocked', { status });
   }
 
-  var viewEventHandlers = {
+  const viewEventHandlers = {
     endCall() {
       otHelper.disconnect();
       const url = window.location.origin.concat('/thanks');
       window.location.href = url;
-
     },
     startArchiving(evt) {
       sendArchivingOperation((evt.detail && evt.detail.operation) || 'startComposite');
@@ -332,7 +269,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
       sendArchivingOperation('stop');
     },
     streamVisibilityChange(evt) {
-      const getStatus = info => {
+      const getStatus = (info) => {
         let status = null;
 
         if (evt.detail.value === 'hidden') {
@@ -354,9 +291,9 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
       }
     },
     buttonClick(evt) {
-      const streamId = evt.detail.streamId;
-      const streamType = evt.detail.streamType;
-      const name = evt.detail.name;
+      const { streamId } = evt.detail;
+      const { streamType } = evt.detail;
+      const { name } = evt.detail;
       const disableAll = !!evt.detail.disableAll;
       const switchStatus = evt.detail.status;
       let buttonInfo = null;
@@ -471,25 +408,25 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         userName,
         token,
         state,
-        roomURI
+        roomURI,
       };
 
       Request.sendLockingOperation(data).then(() => sendSignalLock(state));
-    }
+    },
   };
 
-  var setAudioStatus = switchStatus => {
+  const setAudioStatus = (switchStatus) => {
     otHelper.isPublisherReady && viewEventHandlers.buttonClick({
       detail: {
         streamId: 'publisher',
         name: 'audio',
         disableAll: true,
-        status: switchStatus
-      }
+        status: switchStatus,
+      },
     });
   };
 
-  var sendStatus = (evt, control, enabled) => {
+  const sendStatus = (evt, control, enabled) => {
     let stream = evt.stream || evt.target.stream;
     if (!stream) {
       return;
@@ -503,7 +440,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
     Utils.sendEvent(`roomController:${control}`, {
       id,
       reason: evt.reason,
-      enabled: buttonInfo.enabled
+      enabled: buttonInfo.enabled,
     });
   };
 
@@ -518,14 +455,14 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
     },
     disconnected(evt) {
       Utils.sendEvent('roomController:disconnected', {
-        id: evt.target.stream.streamId
+        id: evt.target.stream.streamId,
       });
     },
     connected(evt) {
       Utils.sendEvent('roomController:connected', {
-        id: evt.target.stream.streamId
+        id: evt.target.stream.streamId,
       });
-    }
+    },
   };
 
   let _allHandlers = {
@@ -554,7 +491,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         // session. For streams published by your own client, the Publisher object
         // dispatches a streamCreated event. For a code example and more details,
         // see StreamEvent.
-        const stream = evt.stream;
+        const { stream } = evt;
         // SIP call streams have no video.
         const streamVideoType = stream.videoType || 'noVideo';
 
@@ -570,7 +507,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
           stream.name = connectionData.name || '';
         }
 
-        const streamId = stream.streamId;
+        const { streamId } = stream;
         stream.phoneNumber = stream.isSip && stream.name;
         if (stream.isSip) {
           stream.name = 'Invited Participant';
@@ -578,7 +515,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
 
         subscriberStreams[streamId] = {
           stream,
-          buttons: SubscriberButtons(streamVideoType, stream.phoneNumber)
+          buttons: SubscriberButtons(streamVideoType, stream.phoneNumber),
         };
 
         const subOptions = subscriberOptions[streamVideoType];
@@ -589,50 +526,43 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         const subsDOMElem = RoomView.createStreamView(streamId, {
           name: stream.name,
           type: stream.videoType,
-          controlElems: subscriberStreams[streamId].buttons
+          controlElems: subscriberStreams[streamId].buttons,
         });
 
         subOptions.subscribeToVideo = !enterWithVideoDisabled;
 
-        /* Use ResizeSensor instead of mutationObserver
-        // We want to observe the container where the actual suscriber will live
-        var subsContainer = LayoutManager.getItemById(streamId);
-        subsContainer && _mutationObserver &&
-          _mutationObserver.observe(subsContainer, { attributes: true });
-        */
-
-        subscriberStreams[streamId].subscriberPromise =
-          otHelper.subscribe(evt.stream, subsDOMElem, subOptions, {}, enableAnnotations)
-            .then(subscriber => {
-              if (streamVideoType === 'screen') {
-                enableAnnotations && Utils.sendEvent('roomController:annotationStarted');
-                const subContainer = subscriber.element.parentElement;
-                Utils.sendEvent('layoutView:itemSelected', {
-                  item: subContainer
-                });
-                return subscriber;
-              }
-
-              Object.keys(_subscriberHandlers).forEach(name => {
-                subscriber.on(name, _subscriberHandlers[name]);
-              });
-              if (enterWithVideoDisabled) {
-                pushSubscriberButton(streamId, 'video', true);
-              }
-
-              new ResizeSensor(subsDOMElem, () => { // eslint-disable-line no-new
-                const subsDimension = {
-                  width: subsDOMElem.clientWidth,
-                  height: subsDOMElem.clientHeight
-                };
-                otHelper.setPreferredResolution(subscriber, null, subsDimension, null, null);
-              });
-
-              sendVideoEvent(evt.stream);
-              return subscriber;
-            }, error => {
-              debug.error(`Error susbscribing new participant. ${error.message}`);
+        subscriberStreams[streamId].subscriberPromise = otHelper.subscribe(
+          evt.stream, subsDOMElem, subOptions, {}, enableAnnotations,
+        ).then((subscriber) => {
+          if (streamVideoType === 'screen') {
+            enableAnnotations && Utils.sendEvent('roomController:annotationStarted');
+            const subContainer = subscriber.element.parentElement;
+            Utils.sendEvent('layoutView:itemSelected', {
+              item: subContainer,
             });
+            return subscriber;
+          }
+
+          Object.keys(_subscriberHandlers).forEach((name) => {
+            subscriber.on(name, _subscriberHandlers[name]);
+          });
+          if (enterWithVideoDisabled) {
+            pushSubscriberButton(streamId, 'video', true);
+          }
+
+          new ResizeSensor(subsDOMElem, () => { // eslint-disable-line no-new
+            const subsDimension = {
+              width: subsDOMElem.clientWidth,
+              height: subsDOMElem.clientHeight,
+            };
+            otHelper.setPreferredResolution(subscriber, null, subsDimension, null, null);
+          });
+
+          sendVideoEvent(evt.stream);
+          return subscriber;
+        }, (error) => {
+          debug.error(`Error susbscribing new participant. ${error.message}`);
+        });
       });
     },
     streamDestroyed(evt) {
@@ -647,7 +577,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
       // For streams published by your own client, the Publisher object
       // dispatches a streamDestroyed event.
       // For a code example and more details, see StreamEvent.
-      const stream = evt.stream;
+      const { stream } = evt;
       if (stream.videoType === 'screen') {
         Utils.sendEvent('roomController:annotationEnded');
       }
@@ -670,7 +600,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
       // Dispatched when an archive recording of the session starts
       Utils.sendEvent('archiving', {
         status: 'started',
-        id: evt.id
+        id: evt.id,
       });
     },
     archiveStopped() {
@@ -684,13 +614,13 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
     'signal:muteAll': function (evt) {
       const statusData = JSON.parse(evt.data);
       const muteAllSwitch = statusData.status;
-      const onlyChangeSwitch = statusData.onlyChangeSwitch;
+      const { onlyChangeSwitch } = statusData;
       // onlyChangeSwitch is true when the iOS app sends a false muteAll signal.
       if (onlyChangeSwitch) {
         return;
       }
 
-      const setNewAudioStatus = (isMuted => {
+      const setNewAudioStatus = ((isMuted) => {
         if (_sharedStatus.roomMuted !== isMuted) {
           return;
         }
@@ -699,23 +629,23 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
 
       if (!otHelper.isMyself(evt.from)) {
         _sharedStatus.roomMuted = muteAllSwitch;
-         setAudioStatus(muteAllSwitch);
-         Utils.sendEvent('roomController:roomMuted', { isJoining: false });
-         RoomView.showConfirmChangeMicStatus(muteAllSwitch).then(setNewAudioStatus);
+        setAudioStatus(muteAllSwitch);
+        Utils.sendEvent('roomController:roomMuted', { isJoining: false });
+        RoomView.showConfirmChangeMicStatus(muteAllSwitch).then(setNewAudioStatus);
       }
     },
     'signal:archives': function (evt) {
       Utils.sendEvent('roomController:archiveUpdates', evt);
-    }
+    },
   };
 
   function showMobileShareUrl() {
     navigator.share({
       title: 'Invite Participant',
-      url: location.href
+      url: location.href,
     })
       .then(() => { console.log('Successful share'); })
-      .catch(error => { console.log('Error sharing', error); });
+      .catch((error) => { console.log('Error sharing', error); });
   }
 
   function addClipBoardFeature(selector) {
@@ -738,21 +668,19 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
 
   function showAddToCallModal() {
     const selector = '.add-to-call-modal';
-    return Modal.show(selector).then(() => {
-      return new Promise(resolve => {
-        const enterButton = document.querySelector(`${selector} button`);
-        enterButton && enterButton.addEventListener('click', function onClicked(event) {
-          event.preventDefault();
-          enterButton.removeEventListener('click', onClicked);
-          if (enterButton.id === 'copyInviteLinkBtn') {
-            addClipBoardFeature(selector);
-          } else {
-            Modal.hide(selector);
-          }
-          resolve();
-        });
+    return Modal.show(selector).then(() => new Promise((resolve) => {
+      const enterButton = document.querySelector(`${selector} button`);
+      enterButton && enterButton.addEventListener('click', function onClicked(event) {
+        event.preventDefault();
+        enterButton.removeEventListener('click', onClicked);
+        if (enterButton.id === 'copyInviteLinkBtn') {
+          addClipBoardFeature(selector);
+        } else {
+          Modal.hide(selector);
+        }
+        resolve();
       });
-    });
+    }));
   }
 
   function getRoomParams() {
@@ -769,7 +697,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
 
     let roomName = '';
     let roomURI = '';
-    const length = pathName.length;
+    const { length } = pathName;
     if (length > 0) {
       roomURI = pathName[length - 1];
     }
@@ -778,12 +706,10 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
     // Recover user identifier
     const params = Utils.parseSearch(document.location.search);
     const usrId = window.userName || params.getFirstValue('userName');
-    resolutionAlgorithm = params.getFirstValue('resolutionAlgorithm');
-    debugPreferredResolution = params.getFirstValue('debugPreferredResolution');
     enableHangoutScroll = params.getFirstValue('enableHangoutScroll') !== undefined;
 
     return PrecallController.showCallSettingsPrompt(roomName, usrId, otHelper)
-      .then(info => {
+      .then((info) => {
         info.roomURI = roomURI;
         RoomView.showRoom();
         RoomView.roomURI = roomURI;
@@ -798,9 +724,9 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
   function getRoomInfo(aRoomParams) {
     return Request
       .getRoomInfo(aRoomParams)
-      .then(aRoomInfo => {
-        if (!(aRoomInfo && aRoomInfo.token && aRoomInfo.sessionId &&
-              aRoomInfo.apiKey && aRoomInfo.username)) {
+      .then((aRoomInfo) => {
+        if (!(aRoomInfo && aRoomInfo.token && aRoomInfo.sessionId
+              && aRoomInfo.apiKey && aRoomInfo.username)) {
           debug.error('Error getRoomParams [', aRoomInfo,
             '] without correct response');
           throw new Error('Error getting room parameters');
@@ -822,7 +748,6 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
     '/js/itemsHandler.js',
     '/js/layoutView.js',
     '/js/layouts.js',
-    '/js/layoutManager.js',
     '/js/roomView.js',
     '/js/roomStatus.js',
     '/js/min/chatController.min.js',
@@ -833,7 +758,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
     '/js/min/feedbackController.min.js',
     '/js/googleAuth.js',
     '/js/min/phoneNumberController.min.js',
-    '/js/vendor/ResizeSensor.js'
+    '/js/vendor/ResizeSensor.js',
   ];
 
   const init = () => {
@@ -844,41 +769,37 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         Utils.addEventsHandlers('precallView:', {
           submit() {
           // Jeff to do: The room logic should go here, not in PrecallController.
-          }
+          },
         });
 
         return PrecallController.init();
       })
-      .then(() => {
-        return LazyLoader.load('/js/helpers/OTHelper.js');
-      })
+      .then(() => LazyLoader.load('/js/helpers/OTHelper.js'))
       .then(() => {
         otHelper = new OTHelper({});
         exports.otHelper = otHelper;
       })
       .then(getRoomParams)
       .then(getRoomInfo)
-      .then(aParams => {
+      .then((aParams) => {
         let loadAnnotations = Promise.resolve();
         if (enableAnnotations) {
-          exports.OTKAnalytics = exports.OTKAnalytics ||
-          (() => {
-            return {
-              addSessionInfo() {},
-              logEvent(a, b) {
-                console.log(a, b); // eslint-disable-line no-console
-              }
-            };
-          });
+          exports.OTKAnalytics = exports.OTKAnalytics
+          || (() => ({
+            addSessionInfo() {},
+            logEvent(a, b) {
+              console.log(a, b); // eslint-disable-line no-console
+            },
+          }));
 
           loadAnnotations = LazyLoader.load([
             'https://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.8.3/underscore-min.js',
-            '/js/vendor/opentok-annotation.js'
+            '/js/vendor/opentok-annotation.js',
           ]);
         }
-        return loadAnnotations.then(() => { return aParams; });
+        return loadAnnotations.then(() => aParams);
       })
-      .then(aParams => {
+      .then((aParams) => {
         RoomView.init(enableHangoutScroll, enableArchiveManager, enableSip);
         // Init this controller before connect to the session
         // to start receiving signals about archives updates
@@ -892,32 +813,30 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
         const sessionInfo = {
           apiKey: aParams.apiKey,
           sessionId: aParams.sessionId,
-          token: aParams.token
+          token: aParams.token,
         };
 
         const connect = otHelper.connect.bind(otHelper, sessionInfo);
 
-        const waitForConnectionCount = () => {
-          return new Promise(resolve => {
-            if (!maxUsersPerRoom) {
-              return resolve();
+        const waitForConnectionCount = () => new Promise((resolve) => {
+          if (!maxUsersPerRoom) {
+            return resolve();
+          }
+          return setTimeout(() => {
+            if (numUsrsInRoom > maxUsersPerRoom) {
+              Utils.sendEvent('roomController:meetingFullError');
+              return;
             }
-            return setTimeout(() => {
-              if (numUsrsInRoom > maxUsersPerRoom) {
-                Utils.sendEvent('roomController:meetingFullError');
-                return;
-              }
-              resolve();
-            }, 500);
-          });
-        };
+            resolve();
+          }, 500);
+        });
 
         RoomView.participantsNumber = 0;
 
         _allHandlers = RoomStatus.init(_allHandlers, { room: _sharedStatus });
 
         if (enableSip && requireGoogleAuth) {
-          GoogleAuth.init(aParams.googleId, aParams.googleHostedDomain, aGoogleAuth => {
+          GoogleAuth.init(aParams.googleId, aParams.googleHostedDomain, (aGoogleAuth) => {
             googleAuth = aGoogleAuth;
             if (googleAuth.isSignedIn.get()) {
               document.body.data('google-signed-in', 'true');
@@ -933,7 +852,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
           .then(() => {
             const publisherElement = RoomView.createStreamView('publisher', {
               name: userName,
-              type: 'publisher'
+              type: 'publisher',
             });
             // If we have all audios disabled, we need to set the button status
             // and don't publish audio
@@ -941,9 +860,9 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
             // Set visual status of button
               sendStatus({
                 stream: {
-                  streamId: 'Publisher'
+                  streamId: 'Publisher',
                 },
-                reason: 'publishAudio'
+                reason: 'publishAudio',
               }, 'audio', false);
               // Don't publish audio
               publisherOptions.publishAudio = false;
@@ -952,7 +871,7 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
             return otHelper.publish(publisherElement, publisherOptions, {}).then(() => {
               setPublisherReady();
               RoomView.showPublisherButtons(publisherOptions);
-            }).catch(errInfo => {
+            }).catch((errInfo) => {
               if (errInfo.error.name === 'OT_CHROME_MICROPHONE_ACQUISITION_ERROR') {
                 Utils.sendEvent('roomController:chromePublisherError');
                 otHelper.disconnect();
@@ -965,14 +884,14 @@ PhoneNumberController, ResizeSensor, maxUsersPerRoom */
             PhoneNumberController.init();
             Utils.sendEvent('roomController:controllersReady');
           })
-          .catch(error => {
+          .catch((error) => {
             debug.error(`Error Connecting to room. ${error.message}`);
           });
       });
   };
 
   const RoomController = {
-    init
+    init,
   };
 
   exports.RoomController = RoomController;
