@@ -1,21 +1,21 @@
 /* global RTCApp */
 
-!(exports => {
+!((exports) => {
   let debug;
 
   const _views = {
     '/room/': {
       mainView: 'RoomController',
       dependencies: [
-        'RoomController'
-      ]
+        'RoomController',
+      ],
     },
     '/': {
       mainView: 'LandingController',
       dependencies: [
-        'LandingController'
-      ]
-    }
+        'LandingController',
+      ],
+    },
   };
 
   function getView() {
@@ -23,12 +23,10 @@
     const numViews = pathViews.length;
     const path = exports.document.location.pathname;
     for (let i = 0; i < numViews; i++) {
-      if (path.startsWith(pathViews[i]) &&
-        _views[pathViews[i]]
+      if (path.startsWith(pathViews[i])
+        && _views[pathViews[i]]
           .dependencies
-          .every(dependency => {
-            return !!exports[dependency];
-          })) {
+          .every((dependency) => !!exports[dependency])) {
         return exports[_views[pathViews[i]].mainView];
       }
     }
@@ -46,12 +44,11 @@
   }
 
   exports.RTCApp = {
-    init
+    init,
   };
 })(this);
 
-
-this.addEventListener('load', function startApp() {
+this.addEventListener('load', () => {
   // Note that since the server forbids loading the content on an iframe this should not execute.
   // But it doesn't hurt either
   if (window.top !== window.self && !window.iframing_allowed) {
@@ -67,7 +64,7 @@ this.addEventListener('load', function startApp() {
       '/shared/js/utils.js',
       '/js/helpers/requests.js',
       '/js/min/roomController.min.js',
-      '/js/landingController.js'
+      '/js/landingController.js',
     ]).then(() => {
       RTCApp.init();
     });
@@ -75,9 +72,9 @@ this.addEventListener('load', function startApp() {
 
   // Allow only https on production
   if (
-    document.location.protocol === 'http:' &&
-    (document.location.hostname.includes('.tokbox.com') ||
-    document.location.hostname.includes('.vonage.com'))
+    document.location.protocol === 'http:'
+    && (document.location.hostname.includes('.tokbox.com')
+    || document.location.hostname.includes('.vonage.com'))
   ) {
     document.location.href = document.location.href.replace(new RegExp('^http:'), 'https:');
   }
